+++ src/extension.ts
@@ -41,10 +41,7 @@ import { registerSchemaCommands } from "./commands/schemas";
 import { registerSearchCommands } from "./commands/search";
 import { registerSupportCommands } from "./commands/support";
 import { registerTopicCommands } from "./commands/topics";
-import {
-  registerDeleteArtifactCommand,
-  registerUploadArtifactCommand,
-} from "./commands/uploadArtifact";
+import { registerArtifactCommand } from "./commands/uploadArtifact";
 import { AUTH_PROVIDER_ID, AUTH_PROVIDER_LABEL, IconNames } from "./constants";
 import { activateMessageViewer } from "./consume";
 import { setExtensionContext } from "./context/extension";
@@ -271,8 +268,7 @@ async function _activateExtension(
     ...registerDocumentCommands(),
     ...registerSearchCommands(),
     ...registerFlinkArtifactCommands(),
-    registerUploadArtifactCommand(),
-    registerDeleteArtifactCommand(),
+    ...registerArtifactCommand(),
     ...registerNewResourceViewCommands(),
   ];
   logger.info("Commands registered");
